Move route handlers into private state
Store the route handlers, guards and param functions along with the rest
of the route's private state. There's no more need to expose these as
private or protected methods on the route itself.

This also simplifies the router, since the route can now directly
return its handler.

Note that handlers are no longer called with the route as the thisArg.
For consistency, the router's fallback is no longer called with the
router as the thisArg.