Support base option in state history
Implement an optional base option when creating state history. The
history value will be relative to the base. When changing the history
the full path, including the base, is used to update the browser's
location.

State history now ensures its values always start with a slash, though
you do not need to provide the slash when changing the history.

Fixes #12.