Full coverage (#32)
* Mark dispatchInitial argument as optional

The implementation assigns a default value, so perhaps it should've been
typed as optional. Add a test for this behavior, increasing code
coverage.

* Refactor navstart defer implementation

Extract into a separate function.

Remove obsolete reference to dojo-core issue.

Create a noop function that's ignored by Istanbul. It's not actually
used but we need the function since TypeScript doesn't know we're
assigning cancel and resume in the promise executor.

* Ignore unused else branches in Istanbul

The if-conditions are only present for TypeScript's benefit, in practice
they should always be true. Therefore the else branch cannot be tested.
Ignore them to bump the code coverage numbers.