Enable noImplicitThis (#33)
* Enable noImplicitThis option

Type this where necessary. Remove now unnecessary type casts.

* Assign property defaults in prototype

These properties are always assigned in the initializer, so they
shouldn't be optional in the interface. However that requires default
values to be provided when composing the constructor.

* Expect route params method to return null, or params

Void is ambiguous.

* Refactor route matching

Change the result types to always have offset and values properties.
Return null in case there is no match. This clears up a lot of strict
null check violations.

* Emit a typed navstart event

Target must be specified, but null seems to be a valid value.