Use a generic for the Context (#41)
* Enable customization of the context type

createRouter() can now be passed a context type using generics. This
type propagates to the error event and dispatch() function.

Similarly createRoute() now supports a variant where it can be passed
context and parameters generics. This propagates to the request object
passed to the various handlers.

Fixes #34.

* Expand C binding in tests to Context

* Pass generic when creating route, don't cast

* Remove unnecessary type specifications

* Simplify fallback route test

No need to declare an interface.