Ensure routes are only appended once
Generating links requires the router to walk the route hierarchy back to
the router itself. This means routes should only be appended to other
routes or the router once.

Add a parent accessor to routes. This can be used both to see whether
the route has previously been appended to another route, and to
walk the hierarchy.